Split out tests for new utils files
